% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurt_skew.R
\name{kurt_skew}
\alias{kurt_skew}
\title{Calculate Model Free Implied Skew & Kurtosis}
\usage{
kurt_skew(m, k, u, ic, ip, currcalls, currputs, er, V)
}
\arguments{
\item{m}{scalar number of grid points}

\item{k}{scalar grid limit}

\item{u}{scalar; (1 + k)^(1 / m)}

\item{ic}{vector of exponents for OTM calls}

\item{ip}{vector of exponents for OTM  puts}

\item{currcalls}{vector of OTM call prices}

\item{currputs}{vector of OTM put prices}

\item{er}{scalar; exp(mat * zero_rate)}

\item{V}{scalar; moments_mfiv_bkm * mat}
}
\value{
list of model free implied skew and kurtosis
}
\description{
Calculate model free implied skew and kurtosis based on BKM (2003)
}
\examples{
See README.md in GitHub

}
\references{
Code based on [updated 2021-06-27] Model-Free Implied Measures from Options Data (Data and Code)" found at https://www.vilkov.net/codedata.html
}
