% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OM_Interpolate_IV.R
\name{OM_Interpolate_IV}
\alias{OM_Interpolate_IV}
\title{Calculate Interpolated Implied Volatility Grid}
\usage{
OM_Interpolate_IV(mnes, vol, grid)
}
\arguments{
\item{mnes}{a vector of moneyness (K/S)}

\item{vol}{a vector of implied volatilities}

\item{grid}{a vector of length two containing m,k (500,2)}
}
\value{
list of interpolated iv and moneyness
}
\description{
Calculate interpolated IV grid based on current IV
}
\examples{
See README.md in GitHub

}
\references{
Code based on [updated 2021-06-27] Model-Free Implied Measures from Options Data (Data and Code)" found at https://www.vilkov.net/codedata.html
}
